+++ types/yanzi/motionPlus/uplink.js
@@ -1,39 +1,35 @@
 function consume(event) {
-  const { data } = event;
-  const { resourceType } = data;
+  const { resourceType } = event.data.list[0];
+  const { value } = event.data.list[0];
   const sample = {};
   let topic = "default";
 
-  /*
-  // Foreach ?
-  if (resourceType === "SampleUpState") {
-    sample.upLog = data.upLog;
-    topic = "upLog";
-  }
-  */
-
   if (resourceType === "SampleHumidity") {
-    sample.humidity = data.relativeHumidity; // * 10
+    sample.humidity = value;
     topic = "humidity";
   }
 
   if (resourceType === "SampleTemp") {
-    sample.temperature = data.temperatureK;
+    sample.temperature = Math.round(value * 100) / 1000;
     topic = "temperature";
   }
 
   if (resourceType === "SampleIlluminance") {
-    sample.light = data.illuminance;
+    sample.light = Math.round(value / 100) / 10;
+    const { colorTemperature } = event.data.list[0];
+    if (colorTemperature !== undefined) {
+      sample.colorTemperature = colorTemperature;
+    }
     topic = "light";
   }
 
   if (resourceType === "SampleSoundPressureLevel") {
-    sample.soundPressure = data.soundPressureLevel; // * 100
+    sample.soundPressure = value / 1000;
     topic = "sound";
   }
 
   if (resourceType === "SampleMotion") {
-    sample.motion = data.motion;
+    sample.motion = value;
     if (sample.motion > 0) {
       sample.occupancy = 1;
     } else {